requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
paddleocr>=2.7.3
httpx>=0.27.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
aiofiles>=23.0.0
numpy>=1.24.0
